+++ rasa/dialogue_understanding/generator/llm_command_generator.py
@@ -25,8 +25,8 @@ from rasa.engine.storage.resource import Resource
 from rasa.engine.storage.storage import ModelStorage
 from rasa.shared.core.flows.flow_step import (
     FlowStep,
-    CollectInformationFlowStep,
 )
+from rasa.shared.core.flows.steps.collect import CollectInformationFlowStep
 from rasa.shared.core.flows.flow import Flow
 from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
