+++ rasa/shared/core/flows/steps/constants.py
@@ -1,0 +1,4 @@
+CONTINUE_STEP_PREFIX = "NEXT:"
+START_STEP = "START"
+END_STEP = "END"
+DEFAULT_STEPS = {END_STEP, START_STEP}
