+++ rasa/shared/core/flows/steps/generate_response.py
@@ -1,0 +1,95 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Text, Optional, Dict, Any
+
+from rasa.shared.core.flows.flow_step import FlowStep, structlogger
+from rasa.shared.core.trackers import DialogueStateTracker
+
+from rasa.shared.utils.llm import (
+    DEFAULT_OPENAI_TEMPERATURE,
+    DEFAULT_OPENAI_GENERATE_MODEL_NAME,
+)
+
+DEFAULT_LLM_CONFIG = {
+    "_type": "openai",
+    "request_timeout": 5,
+    "temperature": DEFAULT_OPENAI_TEMPERATURE,
+    "model_name": DEFAULT_OPENAI_GENERATE_MODEL_NAME,
+}
+
+
+@dataclass
+class GenerateResponseFlowStep(FlowStep):
+    """Represents the configuration of a step prompting an LLM."""
+
+    generation_prompt: Text
+    """The prompt template of the flow step."""
+    llm_config: Optional[Dict[Text, Any]] = None
+    """The LLM configuration of the flow step."""
+
+    @classmethod
+    def from_json(cls, flow_step_config: Dict[Text, Any]) -> GenerateResponseFlowStep:
+        """Used to read flow steps from parsed YAML.
+
+        Args:
+            flow_step_config: The parsed YAML as a dictionary.
+
+        Returns:
+            The parsed flow step.
+        """
+        base = super()._from_json(flow_step_config)
+        return GenerateResponseFlowStep(
+            generation_prompt=flow_step_config.get("generation_prompt", ""),
+            llm_config=flow_step_config.get("llm", None),
+            **base.__dict__,
+        )
+
+    def as_json(self) -> Dict[Text, Any]:
+        """Returns the flow step as a dictionary.
+
+        Returns:
+            The flow step as a dictionary.
+        """
+        dump = super().as_json()
+        dump["generation_prompt"] = self.generation_prompt
+        if self.llm_config:
+            dump["llm"] = self.llm_config
+
+        return dump
+
+    def generate(self, tracker: DialogueStateTracker) -> Optional[Text]:
+        """Generates a response for the given tracker.
+
+        Args:
+            tracker: The tracker to generate a response for.
+
+        Returns:
+            The generated response.
+        """
+        from rasa.shared.utils.llm import llm_factory, tracker_as_readable_transcript
+        from jinja2 import Template
+
+        context = {
+            "history": tracker_as_readable_transcript(tracker, max_turns=5),
+            "latest_user_message": tracker.latest_message.text
+            if tracker.latest_message
+            else "",
+        }
+        context.update(tracker.current_slot_values())
+
+        llm = llm_factory(self.llm_config, DEFAULT_LLM_CONFIG)
+        prompt = Template(self.generation_prompt).render(context)
+
+        try:
+            return llm(prompt)
+        except Exception as e:
+            # unfortunately, langchain does not wrap LLM exceptions which means
+            # we have to catch all exceptions here
+            structlogger.error(
+                "flow.generate_step.llm.error", error=e, step=self.id, prompt=prompt
+            )
+            return None
+
+    def default_id_postfix(self) -> str:
+        return "generate"
