+++ rasa/shared/core/flows/steps/link.py
@@ -1,0 +1,44 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Text, Dict, Any
+
+from rasa.shared.core.flows.flow_step import FlowStep
+
+
+@dataclass
+class LinkFlowStep(FlowStep):
+    """Represents the configuration of a link flow step."""
+
+    link: Text
+    """The link of the flow step."""
+
+    @classmethod
+    def from_json(cls, flow_step_config: Dict[Text, Any]) -> LinkFlowStep:
+        """Used to read flow steps from parsed YAML.
+
+        Args:
+            flow_step_config: The parsed YAML as a dictionary.
+
+        Returns:
+            The parsed flow step.
+        """
+        base = super()._from_json(flow_step_config)
+        return LinkFlowStep(
+            link=flow_step_config.get("link", ""),
+            **base.__dict__,
+        )
+
+    def as_json(self) -> Dict[Text, Any]:
+        """Returns the flow step as a dictionary.
+
+        Returns:
+            The flow step as a dictionary.
+        """
+        dump = super().as_json()
+        dump["link"] = self.link
+        return dump
+
+    def default_id_postfix(self) -> str:
+        """Returns the default id postfix of the flow step."""
+        return f"link_{self.link}"
