+++ tests/dialogue_understanding/commands/test_handle_code_change_command.py
@@ -16,11 +16,8 @@ from rasa.dialogue_understanding.stack.frames import (
 )
 from rasa.shared.core.domain import Domain
 from rasa.shared.core.events import SlotSet
-from rasa.shared.core.flows.flow_step import (
-    START_STEP,
-    ContinueFlowStep,
-    END_STEP,
-)
+from rasa.shared.core.flows.steps.constants import START_STEP, END_STEP
+from rasa.shared.core.flows.steps.continuation import ContinueFlowStep
 from rasa.shared.core.flows.flows_list import FlowsList
 from rasa.shared.core.trackers import DialogueStateTracker
 from tests.dialogue_understanding.commands.test_command_processor import (
